Fix to enable cross compiler to build on Mac OS X 10.6+ (details)
In order to get the cross compiler to build on Mac OS X 10.6+ we needed to
do two things. First, we need to prevent warnings from resulting in an
error when building binutils. We do this with the --disable-werror
configure option. Second we need to replace line 73 of
src/gcc/cp/Make-lang.in to prevent some kind of link error. See this post
for more information:

 http://gcc.gnu.org/ml/gcc-patches/2009-08/msg01583.html

Note that you might also need to use a specific native compiler to build
the cross-compiler. See the setup guide.